refactor: ♻️ refactor files and separate settings, definitions, and main plugin
