+++ src/lib.ts
@@ -1,0 +1,110 @@
+export interface PluginSettings {
+  baseFolderName: string;
+  apiToken: string | null;
+  lastUpdated: string | null;
+  autoSync: boolean;
+  highlightSortOldestToNewest: boolean;
+  highlightSortByLocation: boolean;
+  highlightDiscard: boolean;
+  syncNotesOnly: boolean;
+  colonSubstitute: string;
+  logFile: boolean;
+  logFileName: string;
+  frontMatter: boolean;
+  frontMatterTemplate: string;
+  headerTemplate: string;
+  highlightTemplate: string;
+  useSlugify: boolean;
+  slugifySeparator: string;
+  slugifyLowercase: boolean;
+  deduplicateFiles: boolean;
+  deduplicateProperty: string;
+  deleteDuplicates: boolean;
+  protectFrontmatter: boolean;
+  protectedFields: string;
+  updateFrontmatter: boolean;
+}
+
+export const DEFAULT_SETTINGS: PluginSettings = {
+  baseFolderName: 'Readwise',
+  apiToken: null,
+  lastUpdated: null,
+  autoSync: true,
+  highlightSortOldestToNewest: true,
+  highlightSortByLocation: true,
+  highlightDiscard: false,
+  syncNotesOnly: false,
+  colonSubstitute: '-',
+  logFile: true,
+  logFileName: 'Sync.md',
+  frontMatter: false,
+  frontMatterTemplate: `---
+id: {{ id }}
+created: {{ created }}
+updated: {{ updated }}
+title: {{ title }}
+author: {{ author }}
+---
+`,
+  headerTemplate: `
+%%
+ID: {{ id }}
+Updated: {{ updated }}
+%%
+
+![]( {{ cover_image_url }})
+
+# About
+Title: [[{{ sanitized_title }}]]
+Authors: {{ authorStr }}
+Category: #{{ category }}
+{%- if tags %}
+Tags: {{ tags }}
+{%- endif %}
+Number of Highlights: =={{ num_highlights }}==
+Readwise URL: {{ highlights_url }}
+{%- if source_url %}
+Source URL: {{ source_url }}
+{%- endif %}
+Date: [[{{ created }}]]
+Last Highlighted: *{{ last_highlight_at }}*
+{%- if summary %}
+Summary: {{ summary }}
+{%- endif %}
+
+---
+
+{%- if document_note %}
+# Document Note
+
+{{ document_note }}
+{%- endif %}
+
+# Highlights
+
+`,
+  highlightTemplate: `{{ text }}{%- if category == 'books' %} ([{{ location }}]({{ location_url }})){%- endif %}{%- if color %} %% Color: {{ color }} %%{%- endif %} ^{{id}}{%- if note %}
+
+Note: {{ note }}
+{%- endif %}{%- if tags %}
+
+Tags: {{ tags }}
+{%- endif %}{%- if url %}
+
+[View Highlight]({{ url }})
+{%- endif %}
+
+---
+`,
+  useSlugify: false,
+  slugifySeparator: '-',
+  slugifyLowercase: true,
+  deduplicateFiles: false,
+  deduplicateProperty: 'uri',
+  deleteDuplicates: true,
+  protectFrontmatter: false,
+  protectedFields: 'connections\nstatus\ntags',
+  updateFrontmatter: true,
+};
+
+export const FRONTMATTER_TO_ESCAPE = ['title', 'sanitized_title', 'author', 'authorStr'];
